move back tests to TestSpec
